feat(api): complete portfolio value

+++ src/rest/schemas/responses/portfolio.py
@@ -1,5 +1,6 @@
+from decimal import Decimal
 from pydantic import BaseModel
 
 
 class PortfolioValueResponse(BaseModel):
-    value: float
+    value: Decimal
